Refactor signer dispatching related structs into their own modules under crypto::signing.
