+++ src/commons/crypto/signing/dispatch/signerrouter.rs
@@ -1,204 +1,34 @@
-//! Support for signing mft, crl, certificates, roas..
-//! Common objects for TAs and CAs
 use std::{
+    path::Path,
     sync::{Arc, RwLock},
-    {convert::TryFrom, path::Path},
 };
 
-use bytes::Bytes;
-
-use rpki::{
-    repository::{
-        cert::{Cert, KeyUsage, Overclaim, TbsCert},
-        crl::{Crl, CrlEntry, TbsCertList},
-        crypto::{
-            signer::KeyError, DigestAlgorithm, KeyIdentifier, PublicKey, PublicKeyFormat, Signature,
-            SignatureAlgorithm, Signer, SigningError,
-        },
-        csr::Csr,
-        manifest::{FileAndHash, Manifest, ManifestContent},
-        roa::{Roa, RoaBuilder},
-        rta,
-        sigobj::SignedObjectBuilder,
-        x509::{Name, Serial, Time, Validity},
-    },
-    uri,
+use rpki::repository::crypto::{
+    signer::KeyError, KeyIdentifier, PublicKey, PublicKeyFormat, Signature, SignatureAlgorithm, Signer, SigningError,
 };
 
-use crate::{
-    commons::{
-        api::{IssuedCert, RcvdCert, ReplacedObject, RepoInfo, RequestResourceLimit, ResourceSet},
-        crypto::{
-            self,
-            signers::{error::SignerError, softsigner::OpenSslSigner},
-            CryptoResult,
-        },
-        error::Error,
-        util::AllowedUri,
-        KrillResult,
-    },
-    daemon::ca::CertifiedKey,
-};
-
-#[cfg(feature = "hsm")]
 use crate::commons::{
-    api::Handle,
-    crypto::signers::{kmip::KmipSigner, signerinfo::SignerMapper},
+    crypto::{
+        dispatch::signerprovider::SignerProvider,
+        signers::{error::SignerError, softsigner::OpenSslSigner},
+    },
+    KrillResult,
 };
 
 #[cfg(feature = "hsm")]
 use std::collections::HashMap;
 
-//------------ SignerProvider ------------------------------------------------
-
-/// Dispatchers Signer requests to a particular implementation of the Signer trait.
-///
-/// Named and modelled after the similar AuthProvider concept that already exists in Krill.
-#[allow(dead_code)] // Needed as we currently only ever construct one variant
-#[derive(Clone, Debug)]
-enum SignerProvider {
-    OpenSsl(OpenSslSigner),
-
-    #[cfg(feature = "hsm")]
-    Kmip(KmipSigner),
-}
-
-impl SignerProvider {
-    pub fn supports_random(&self) -> bool {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.supports_random(),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.supports_random(),
-        }
-    }
-
-    #[cfg(feature = "hsm")]
-    pub fn create_registration_key(&mut self) -> Result<(PublicKey, String), SignerError> {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.create_registration_key(),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.create_registration_key(),
-        }
-    }
-
-    #[cfg(feature = "hsm")]
-    pub fn sign_registration_challenge<D: AsRef<[u8]> + ?Sized>(
-        &self,
-        internal_key_id: String,
-        challenge: &D,
-    ) -> Result<Signature, SignerError> {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.sign_registration_challenge(internal_key_id, challenge),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.sign_registration_challenge(internal_key_id, challenge),
-        }
-    }
-
-    #[cfg(feature = "hsm")]
-    pub fn get_handle(&self) -> Option<Handle> {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.get_handle(),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.get_handle(),
-        }
-    }
-
-    #[cfg(feature = "hsm")]
-    pub fn set_handle(&mut self, handle: Handle) {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.set_handle(handle),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.set_handle(handle),
-        }
-    }
-
-    #[cfg(feature = "hsm")]
-    pub fn get_name(&self) -> &str {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.get_name(),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.get_name(),
-        }
-    }
-
-    #[cfg(feature = "hsm")]
-    pub fn get_info(&self) -> Option<String> {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.get_info(),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.get_info(),
-        }
-    }
-}
-
-impl Signer for SignerProvider {
-    type KeyId = KeyIdentifier;
-
-    type Error = SignerError;
-
-    fn create_key(&mut self, algorithm: PublicKeyFormat) -> Result<Self::KeyId, Self::Error> {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.create_key(algorithm),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.create_key(algorithm),
-        }
-    }
-
-    fn get_key_info(&self, key: &Self::KeyId) -> Result<PublicKey, KeyError<Self::Error>> {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.get_key_info(key),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.get_key_info(key),
-        }
-    }
-
-    fn destroy_key(&mut self, key: &Self::KeyId) -> Result<(), KeyError<Self::Error>> {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.destroy_key(key),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.destroy_key(key),
-        }
-    }
-
-    fn sign<D: AsRef<[u8]> + ?Sized>(
-        &self,
-        key: &Self::KeyId,
-        algorithm: SignatureAlgorithm,
-        data: &D,
-    ) -> Result<Signature, SigningError<Self::Error>> {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.sign(key, algorithm, data),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.sign(key, algorithm, data),
-        }
-    }
-
-    fn sign_one_off<D: AsRef<[u8]> + ?Sized>(
-        &self,
-        algorithm: SignatureAlgorithm,
-        data: &D,
-    ) -> Result<(Signature, PublicKey), Self::Error> {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.sign_one_off(algorithm, data),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.sign_one_off(algorithm, data),
-        }
-    }
-
-    fn rand(&self, target: &mut [u8]) -> Result<(), Self::Error> {
-        match self {
-            SignerProvider::OpenSsl(signer) => signer.rand(target),
-            #[cfg(feature = "hsm")]
-            SignerProvider::Kmip(signer) => signer.rand(target),
-        }
-    }
-}
+#[cfg(feature = "hsm")]
+use crate::commons::{
+    api::Handle,
+    crypto::{dispatch::signerinfo::SignerMapper, signers::kmip::KmipSigner},
+};
 
 //------------ SignerRouter --------------------------------------------------
 
 /// Manages multiple Signers and routes requests to the appropriate Signer.
 #[derive(Clone, Debug)]
-struct SignerRouter {
+pub struct SignerRouter {
     // TODO: Remove the [RwLock] around the [SignerProvider] once the [Signer] trait is modified to delegate locking to
     // the implementation and is thus able to drop use of the `&mut self` argument which is what cause us to need the
     // `RwLock`. See: https://github.com/NLnetLabs/rpki-rs/issues/161
@@ -483,12 +313,12 @@ impl SignerRouter {
 // pattern with as much or as little locking internally at the finest level of granularity possible.
 // Update: https://github.com/NLnetLabs/rpki-rs/pull/162 removes the &mut.
 impl SignerRouter {
-    fn create_key_minimally_locking(&self, algorithm: PublicKeyFormat) -> Result<KeyIdentifier, SignerError> {
+    pub fn create_key_minimally_locking(&self, algorithm: PublicKeyFormat) -> Result<KeyIdentifier, SignerError> {
         self.register_pending_signers()?;
         self.default_signer.write().unwrap().create_key(algorithm)
     }
 
-    fn destroy_key_minimally_locking(&self, key_id: &KeyIdentifier) -> Result<(), KeyError<SignerError>> {
+    pub fn destroy_key_minimally_locking(&self, key_id: &KeyIdentifier) -> Result<(), KeyError<SignerError>> {
         self.register_pending_signers()?;
         self.get_signer_for_key(key_id)?.write().unwrap().destroy_key(key_id)
     }
@@ -686,351 +516,5 @@ impl Signer for SignerRouter {
         } else {
             self.rand_fallback_signer.read().unwrap().rand(target)
         }
-    }
-}
-
-//------------ KrillSigner ---------------------------------------------------
-
-/// High level signing interface between Krill and the Signer backends.
-///
-/// KrillSigner:
-///   - Is configured via the Krill configuration file.
-///   - Maps Result<SignerError> to KrillResult.
-///   - Directs signers to use the RPKI standard key format (RSA).
-///   - Directs signers to use the RPKI standard signature algorithm (RSA PKCS #1 v1.5 with SHA-256).
-///   - Offers a higher level interface than the Signer trait.
-#[derive(Clone, Debug)]
-pub struct KrillSigner {
-    router: SignerRouter,
-}
-
-impl KrillSigner {
-    pub fn build(work_dir: &Path) -> KrillResult<Self> {
-        Ok(KrillSigner {
-            router: SignerRouter::build(work_dir)?,
-        })
-    }
-
-    pub fn create_key(&self) -> CryptoResult<KeyIdentifier> {
-        self.router
-            .create_key_minimally_locking(PublicKeyFormat::Rsa)
-            .map_err(crypto::Error::signer)
-    }
-
-    pub fn destroy_key(&self, key_id: &KeyIdentifier) -> CryptoResult<()> {
-        self.router
-            .destroy_key_minimally_locking(key_id)
-            .map_err(crypto::Error::key_error)
-    }
-
-    pub fn get_key_info(&self, key_id: &KeyIdentifier) -> CryptoResult<PublicKey> {
-        self.router.get_key_info(key_id).map_err(crypto::Error::key_error)
-    }
-
-    pub fn random_serial(&self) -> CryptoResult<Serial> {
-        Serial::random(&self.router).map_err(crypto::Error::signer)
-    }
-
-    pub fn sign<D: AsRef<[u8]> + ?Sized>(&self, key_id: &KeyIdentifier, data: &D) -> CryptoResult<Signature> {
-        self.router
-            .sign(key_id, SignatureAlgorithm::default(), data)
-            .map_err(crypto::Error::signing)
-    }
-
-    pub fn sign_one_off<D: AsRef<[u8]> + ?Sized>(&self, data: &D) -> CryptoResult<(Signature, PublicKey)> {
-        self.router
-            .sign_one_off(SignatureAlgorithm::default(), data)
-            .map_err(crypto::Error::signer)
-    }
-
-    pub fn sign_csr(&self, base_repo: &RepoInfo, name_space: &str, key: &KeyIdentifier) -> CryptoResult<Csr> {
-        let pub_key = self.router.get_key_info(key).map_err(crypto::Error::key_error)?;
-        let enc = Csr::construct(
-            &self.router,
-            key,
-            &base_repo.ca_repository(name_space).join(&[]).unwrap(), // force trailing slash
-            &base_repo.rpki_manifest(name_space, &pub_key.key_identifier()),
-            Some(&base_repo.rpki_notify()),
-        )
-        .map_err(crypto::Error::signing)?;
-        Ok(Csr::decode(enc.as_slice())?)
-    }
-
-    pub fn sign_cert(&self, tbs: TbsCert, key_id: &KeyIdentifier) -> CryptoResult<Cert> {
-        tbs.into_cert(&self.router, key_id).map_err(crypto::Error::signing)
-    }
-
-    pub fn sign_crl(&self, tbs: TbsCertList<Vec<CrlEntry>>, key_id: &KeyIdentifier) -> CryptoResult<Crl> {
-        tbs.into_crl(&self.router, key_id).map_err(crypto::Error::signing)
-    }
-
-    pub fn sign_manifest(
-        &self,
-        content: ManifestContent,
-        builder: SignedObjectBuilder,
-        key_id: &KeyIdentifier,
-    ) -> CryptoResult<Manifest> {
-        content
-            .into_manifest(builder, &self.router, key_id)
-            .map_err(crypto::Error::signing)
-    }
-
-    pub fn sign_roa(
-        &self,
-        roa_builder: RoaBuilder,
-        object_builder: SignedObjectBuilder,
-        key_id: &KeyIdentifier,
-    ) -> CryptoResult<Roa> {
-        roa_builder
-            .finalize(object_builder, &self.router, key_id)
-            .map_err(crypto::Error::signing)
-    }
-
-    pub fn sign_rta(&self, rta_builder: &mut rta::RtaBuilder, ee: Cert) -> CryptoResult<()> {
-        let key = ee.subject_key_identifier();
-        rta_builder.push_cert(ee);
-        rta_builder
-            .sign(&self.router, &key, None, None)
-            .map_err(crypto::Error::signing)
-    }
-}
-
-//------------ CsrInfo -------------------------------------------------------
-
-pub type CaRepository = uri::Rsync;
-pub type RpkiManifest = uri::Rsync;
-pub type RpkiNotify = uri::Https;
-
-pub struct CsrInfo {
-    ca_repository: CaRepository,
-    rpki_manifest: RpkiManifest,
-    rpki_notify: Option<RpkiNotify>,
-    key: PublicKey,
-}
-
-impl CsrInfo {
-    pub fn new(
-        ca_repository: CaRepository,
-        rpki_manifest: RpkiManifest,
-        rpki_notify: Option<RpkiNotify>,
-        key: PublicKey,
-    ) -> Self {
-        CsrInfo {
-            ca_repository,
-            rpki_manifest,
-            rpki_notify,
-            key,
-        }
-    }
-
-    pub fn global_uris(&self) -> bool {
-        self.ca_repository.seems_global_uri()
-            && self.rpki_manifest.seems_global_uri()
-            && self
-                .rpki_notify
-                .as_ref()
-                .map(|uri| uri.seems_global_uri())
-                .unwrap_or_else(|| true)
-    }
-
-    pub fn unpack(self) -> (CaRepository, RpkiManifest, Option<RpkiNotify>, PublicKey) {
-        (self.ca_repository, self.rpki_manifest, self.rpki_notify, self.key)
-    }
-
-    pub fn key_id(&self) -> KeyIdentifier {
-        self.key.key_identifier()
-    }
-}
-
-impl TryFrom<&Csr> for CsrInfo {
-    type Error = Error;
-
-    fn try_from(csr: &Csr) -> KrillResult<CsrInfo> {
-        csr.validate().map_err(|_| Error::invalid_csr("invalid signature"))?;
-        let ca_repository = csr
-            .ca_repository()
-            .cloned()
-            .ok_or_else(|| Error::invalid_csr("missing ca repository"))?;
-        let rpki_manifest = csr
-            .rpki_manifest()
-            .cloned()
-            .ok_or_else(|| Error::invalid_csr("missing rpki manifest"))?;
-        let rpki_notify = csr.rpki_notify().cloned();
-        let key = csr.public_key().clone();
-        Ok(CsrInfo {
-            ca_repository,
-            rpki_manifest,
-            rpki_notify,
-            key,
-        })
-    }
-}
-
-impl From<&Cert> for CsrInfo {
-    fn from(issued: &Cert) -> Self {
-        let ca_repository = issued.ca_repository().cloned().unwrap();
-        let rpki_manifest = issued.rpki_manifest().cloned().unwrap();
-        let rpki_notify = issued.rpki_notify().cloned();
-        let key = issued.subject_public_key_info().clone();
-        CsrInfo {
-            ca_repository,
-            rpki_manifest,
-            rpki_notify,
-            key,
-        }
-    }
-}
-
-//------------ CaSignSupport -------------------------------------------------
-
-/// Support signing by CAs
-pub struct SignSupport;
-
-impl SignSupport {
-    /// Create an IssuedCert
-    pub fn make_issued_cert(
-        csr: CsrInfo,
-        resources: &ResourceSet,
-        limit: RequestResourceLimit,
-        replaces: Option<ReplacedObject>,
-        signing_key: &CertifiedKey,
-        weeks: i64,
-        signer: &KrillSigner,
-    ) -> KrillResult<IssuedCert> {
-        let signing_cert = signing_key.incoming_cert();
-        let resources = resources.apply_limit(&limit)?;
-        if !signing_cert.resources().contains(&resources) {
-            return Err(Error::MissingResources);
-        }
-
-        let validity = Self::sign_validity_weeks(weeks);
-        let request = CertRequest::Ca(csr, validity);
-
-        let tbs = Self::make_tbs_cert(&resources, signing_cert, request, signer)?;
-        let cert = signer.sign_cert(tbs, signing_key.key_id())?;
-
-        let cert_uri = signing_cert.uri_for_object(&cert);
-
-        Ok(IssuedCert::new(cert_uri, limit, resources, cert, replaces))
-    }
-
-    /// Create an EE certificate for use in ResourceTaggedAttestations.
-    /// Note that for RPKI signed objects such as ROAs and Manifests, the
-    /// EE certificate is created by the rpki.rs library instead.
-    pub fn make_rta_ee_cert(
-        resources: &ResourceSet,
-        signing_key: &CertifiedKey,
-        validity: Validity,
-        pub_key: PublicKey,
-        signer: &KrillSigner,
-    ) -> KrillResult<Cert> {
-        let signing_cert = signing_key.incoming_cert();
-        let request = CertRequest::Ee(pub_key, validity);
-        let tbs = Self::make_tbs_cert(resources, signing_cert, request, signer)?;
-
-        let cert = signer.sign_cert(tbs, signing_key.key_id())?;
-        Ok(cert)
-    }
-
-    fn make_tbs_cert(
-        resources: &ResourceSet,
-        signing_cert: &RcvdCert,
-        request: CertRequest,
-        signer: &KrillSigner,
-    ) -> KrillResult<TbsCert> {
-        let serial = signer.random_serial()?;
-        let issuer = signing_cert.cert().subject().clone();
-
-        let validity = match &request {
-            CertRequest::Ca(_, validity) => *validity,
-            CertRequest::Ee(_, validity) => *validity,
-        };
-
-        let pub_key = match &request {
-            CertRequest::Ca(info, _) => info.key.clone(),
-            CertRequest::Ee(key, _) => key.clone(),
-        };
-
-        let subject = Some(Name::from_pub_key(&pub_key));
-
-        let key_usage = match &request {
-            CertRequest::Ca(_, _) => KeyUsage::Ca,
-            CertRequest::Ee(_, _) => KeyUsage::Ee,
-        };
-
-        let overclaim = Overclaim::Refuse;
-
-        let mut cert = TbsCert::new(serial, issuer, validity, subject, pub_key, key_usage, overclaim);
-
-        let asns = resources.to_as_resources();
-        if asns.is_inherited() || !asns.to_blocks().unwrap().is_empty() {
-            cert.set_as_resources(asns);
-        }
-
-        let ipv4 = resources.to_ip_resources_v4();
-        if ipv4.is_inherited() || !ipv4.to_blocks().unwrap().is_empty() {
-            cert.set_v4_resources(ipv4);
-        }
-
-        let ipv6 = resources.to_ip_resources_v6();
-        if ipv6.is_inherited() || !ipv6.to_blocks().unwrap().is_empty() {
-            cert.set_v6_resources(ipv6);
-        }
-
-        cert.set_authority_key_identifier(Some(signing_cert.cert().subject_key_identifier()));
-        cert.set_ca_issuer(Some(signing_cert.uri().clone()));
-        cert.set_crl_uri(Some(signing_cert.crl_uri()));
-
-        match request {
-            CertRequest::Ca(csr, _) => {
-                let (ca_repository, rpki_manifest, rpki_notify, _pub_key) = csr.unpack();
-                cert.set_basic_ca(Some(true));
-                cert.set_ca_repository(Some(ca_repository));
-                cert.set_rpki_manifest(Some(rpki_manifest));
-                cert.set_rpki_notify(rpki_notify);
-            }
-            CertRequest::Ee(_, _) => {
-                // cert.set_signed_object() ??
-            }
-        }
-
-        Ok(cert)
-    }
-
-    /// Returns a validity period from 5 minutes ago (in case of NTP mess-up), to
-    /// X weeks from now.
-    pub fn sign_validity_weeks(weeks: i64) -> Validity {
-        let from = Time::five_minutes_ago();
-        let until = Time::now() + chrono::Duration::weeks(weeks);
-        Validity::new(from, until)
-    }
-
-    pub fn sign_validity_days(days: i64) -> Validity {
-        let from = Time::five_minutes_ago();
-        let until = Time::now() + chrono::Duration::days(days);
-        Validity::new(from, until)
-    }
-}
-
-#[allow(clippy::large_enum_variant)]
-enum CertRequest {
-    Ca(CsrInfo, Validity),
-    Ee(PublicKey, Validity),
-}
-
-trait ManifestEntry {
-    fn mft_bytes(&self) -> Bytes;
-    fn mft_hash(&self) -> Bytes {
-        let digest = DigestAlgorithm::default().digest(self.mft_bytes().as_ref());
-        Bytes::copy_from_slice(digest.as_ref())
-    }
-    fn mft_entry(&self, name: &str) -> FileAndHash<Bytes, Bytes> {
-        FileAndHash::new(Bytes::copy_from_slice(name.as_bytes()), self.mft_hash())
-    }
-}
-
-impl ManifestEntry for Crl {
-    fn mft_bytes(&self) -> Bytes {
-        self.to_captured().into_bytes()
     }
 }
