+++ src/commons/crypto/signing/signers/mod.rs
@@ -1,7 +1,8 @@
 pub mod error;
+
 #[cfg(feature = "hsm")]
 pub mod kmip;
-#[cfg(feature = "hsm")]
-pub mod signerinfo;
+
 pub mod softsigner;
+
 pub mod util;
